# docs/m2-api-openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "M2 API",
    "description": "Centralized AI communications platform",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    { "url": "http://localhost:3000", "description": "Local dev" }
  ],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" }
    },
    "schemas": {
...
      "CreateChannelDto": {
        "type": "object",
        "properties": {
          "type": { "type": "string", "enum": ["WHATSAPP", "INSTAGRAM", "MESSENGER", "TELEGRAM", "WEBCHAT"], "description": "Channel type" },
          "provider": { "type": "string", "enum": ["META", "EVOLUTION"], "description": "Channel provider (META for Cloud API, EVOLUTION for Evolution API)", "default": "META" },
          "name": { "type": "string", "description": "Channel display name", "maxLength": 255 },
          "credentials": { "type": "object", "description": "Channel credentials (provider-specific)", "additionalProperties": true },
          "isActive": { "type": "boolean", "description": "Whether the channel is active (enabled for use). Use DELETE endpoint to remove it.", "default": true }
        },
        "required": ["type", "name", "credentials"]
      },
...
    }
  },
  "paths": {
...
    "/api/channels": {
      "post": {
        "operationId": "ChannelsController_create",
        "summary": "Create a new channel",
        "tags": ["Channels"],
        "security": [{ "bearer": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateChannelDto" } } } },
        "responses": { "201": { "description": "Channel created successfully" } }
      },
      "get": {
        "operationId": "ChannelsController_findAll",
        "summary": "List all channels (paginated, excludes soft-deleted)",
        "tags": ["Channels"],
        "security": [{ "bearer": [] }],
        "parameters": [
          { "name": "page", "required": false, "in": "query", "schema": { "minimum": 1, "default": 1, "type": "number" } },
          { "name": "limit", "required": false, "in": "query", "schema": { "minimum": 1, "maximum": 100, "default": 20, "type": "number" } },
          { "name": "sortBy", "required": false, "in": "query", "schema": { "type": "string" } },
          { "name": "sortOrder", "required": false, "in": "query", "schema": { "default": "desc", "type": "string", "enum": ["asc", "desc"] } }
        ],
        "responses": { "200": { "description": "Channels retrieved successfully" } }
      }
    },
...
    "/api/channels/{channelId}/evolution/connect": {
      "post": {
        "operationId": "EvolutionController_connect",
        "summary": "Create Evolution instance (or reuse existing) and return QR/state",
        "tags": ["Evolution"],
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "channelId", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": { "201": { "description": "Instance ensured; webhook registered; QR or connection state returned" }, "400": { "description": "Channel is not an Evolution provider" } }
      }
    },
    "/api/channels/{channelId}/evolution/qr": {
      "get": {
        "operationId": "EvolutionController_getQrCode",
        "summary": "Get current QR code for WhatsApp connection",
        "tags": ["Evolution"],
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "channelId", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "QR code data returned" }, "400": { "description": "Channel is not an Evolution provider" } }
      }
    },
    "/api/channels/{channelId}/evolution/status": {
      "get": {
        "operationId": "EvolutionController_getStatus",
        "summary": "Get Evolution instance connection status",
        "tags": ["Evolution"],
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "channelId", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Connection status returned" }, "400": { "description": "Channel is not an Evolution provider" } }
      }
    },
    "/api/channels/{channelId}/evolution/sync-contacts": {
      "post": {
        "operationId": "EvolutionController_syncContacts",
        "summary": "Sync contacts from paired device phonebook",
        "tags": ["Evolution"],
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "channelId", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Contacts synchronized" }, "400": { "description": "Channel is not an Evolution provider" } }
      }
    },
...
  }
}